Basic UI completed.  All components fully functional.  Data updates automatically after each CRUD operation.
